web/requirements.txt
flask
psycopg2
gunicorn
Pillow
